Bug 1352907 - Simplify extension system for performance
The extension system spends a lot of time on concerns we don't have.
We only care about extensions in the extension directory -- nothing else is
used. Supporting all the variations of how extensions can be loaded causes it to
be slow.

It also causes loading regular modules to be slow, because it adds
and @INC hook for every extension.

This commit uses a single @INC hook and makes some assumptions
about how extensions are organized. As a result it is much faster.