Bug 1357775 - the needinfo dropdown for the triage owner should display the name and email address when selected
Turns out I missed the logic for displaying the real name beside the needinfo
role selector. That logic is scary and weird and I would've replaced it had I
noticed it then.

This patch removes the template code that generated JS, and instead puts
populates a data-identity attribute with the real name (or login name).

I added a convenience method to Bugzilla::User for this, which can later be
upstreamed.