rewrite large portions of this
rather than managing a big gpg homedir that could contain millions of
users, I make it so each user has their own homedir. The homedir is
based on the user id and a hash of their public key. Calling
$user->gpg_homedir() causes it to be created if it doesn't exist.

Some other script will need to run the periodically free up space, but
this frees us from a lot of management tasks.

Now for the bad part: when I try to use this, it seems like gpg hangs.

:(