Bug 1440239 - PhabBugz - Assign a secure-revision.
Phabricator's Herald has a "Secure Email" rule and Phabricator sends a
restricted email notification if a revision belongs to the newly created
"secure-revision" group project.

With this patch Bugzilla will automatically assign a secure revision to
the "secure-revision" group project.

Additionally:
The get_project_phid is modified and is now caching the response in
memcached.