Bug 1440328 - Obfuscate mentor email for users that aren't logged in
Throughout bugzilla we normally remove the host of an email address
when displaying the page to a user that is not authenticated. This
prevents spam bots and such from scrapping people's emails.

This was not being done for mentor email addresses in the activity
stream and bug history. See Bug 1440328.

This commit properly filters the mentor email addresses in all places
where we commonly filter emails.