PhabBugz: Fix the unaccepted revision comment
Summary:
Currently, we're sending the "User removed from revision" comment when
the Accept flag has been removed from a revision. This could happen in a
number of use cases:
1. The reviewer resigned from being a reviewer.
2. The reviewer has been removed from reviewers list.
3. Someone (author, reviewer) removed acceptance of the revision by changing
   its status to "Needs Review".

This patch is sending a "flag is deactivated" or "reviewer removed from
revision" depending on the current status of the reviewer.

Test Plan:
Accept a revision.
Change the revision to "Needs Review".
Check the bug comment.
Accept the revision.
Remove the user from reviewers list.
Check the bug comment.

Reviewers: dkl

Bug #: 1441063

Differential Revision: https://phabricator.services.mozilla.com/D809