Add data-no-csrf to 'safe' forms
This is to fix https://bugzilla.mozilla.org/show_bug.cgi?id=1446431
The data-no-csrf attribute is used to signify that a form is 'safe' (ie
doesnt actually make any permanent changes) and so doesnt need an
anti-csrf token.
If any of the forms I've changed do actually make permanent changes then
please flag them to me :)
They may be other forms that need this attribute but hopefully this is a
good start.