Bug 1438205 - Preserve comments in progress across page reloads
Comments being typed will be saved in Local Storage for a week.
When the user somehow reloads the show bug page, the comment box
will be filled in with what was last saved.

This also fixes the bug where you type up a comment, go to change
the bug details, hit cancel, and ulitmately lose your comment in
progress because of the reloading nature of the cancel button.