Bug 1456878 - Fix blockquotes and reply to button for markdown comments
For plaintext comments the comment as fetched from the containing html
element can be quoted. But, for markdown comments it is vital that we
quote the raw markdown. Since the html element contains the rendered
markdown, we fetch the raw text from the BMO api.

This commit also adds some new modern styling for blockquotes.