Bug 1540715 - Fix offset of historical dates in secbugs report
`FILTER time` from Bugzilla's template helpers was inconsistent on
how it rendered the dates of the columns (due to different system
timezones that change based on where the code is being run).

This patch sets the expected timezone for the report and manually
formats the time (since FILTER time seems to not respect the date's
timezone).