Bug 1550145 - Add 'forget_after_date' field to profiles table
Adds a datetime `forget_after_date` field to the profiles table.
This field will be used when determining which unresponsive inactive users to remove periodically.

This field will be normally set to 30 days after BMO notifies that a user has been "inactive"
for a very long time, thus, giving the user 30 days to respond to prevent being forgotten.

Another benefit of this datetime field is that we can choose to send reminder emails as their
forget-date approaches (we may choose not to do this for other reasons).

`forget_after_date`:
- If null, the user should not be forgotten.
- If set, the user can & should be forgotten at anytime after the set date.
- This field should be unset anytime the user logs in or specifically requests to not be forgotten.