no bug - remove warning "Odd number of elements"
The callback jwt_claims is called in list context.
Without an explicit return, when the if (!...) does not match it will
return a single false item. This causes the "Odd number of elements..."
perl warning.

Adding a `return` will prevent the warning and is also probably the
intent of the code.