Bug 1825961: sync revision status with `approval-mozilla-{repo}` flag
When processing revision attachments changes to bugs in phab-bot
feed reader, check if the attachment belongs to an uplift revision
and sync the approval flag statuses as appropriate. To do this
we make the `set_approval_flag_statuses` function more generic
by accepting a flag status instead of hard-coding the `?` status.
We also move the call to update the attachment to the caller.