Bug 1861776: improve usability of the uplift request form comment
Add an array which contains the correct order of uplift request form
fields to ensure they are always presented in the correct order in
the comment. Add the repo's short name to the header so it is more
obvious which request the form corresponds to. Use a smaller header
size to more closely align with the existing Bugzilla uplift request
form.