Remove padding from almost all cells.
We keep it on the cell that has the right-hand border, this keeps some padding for all cell heights; and we keep the border.
We leave it on that cell in particular so we don't need a 'style=' on any other cell.

We also remove the alternating style info for the last columns since we'll be removing those