Bug 1985621: trigger uplift changes when hash of last form changes
After bug 1982619 is deployed to Lando, the uplift request form
will be updated via the Conduit API. When the form is updated via
the API, no feed story is produced, and the transactions view does
not provide enough detail to determine if the uplift form has changed.

Add a new table to track a hash of the contents of the uplift request
form for each revision. When processing a revision change, if the
uplift request form is non-empty, compare against the previously
stored hash and perform the uplift processing steps if the hash has
changed, then store the new hash for the revision.

After processing the form, if the story text indicates the event was
triggered from an update of the uplift request form via the Phabricator
form, we can exit early as there are no more actions to be taken.